// src/app/models/vehicle.ts
export class Vehicle {
  vehicleMovementId: number;
  year: number;
  make: string;
  model: string;
  trim: string;
  mileage: number;
  vin: string;
  color: string;
  location: string;
  zipcode: number;
  askingPrice: number;
  kbb: number;
  crRating: number;
  inspected: boolean;
  imageFile: string;
  bidder: string;
  bidAmount: number;
  bidderImageUrl: string;
}